Further refactor of pytorch tests to clean up code and prepare for upstreamable patch.
